Fix shell error (#86)
* Fix shell error

* Fix warning for line_long

* Modifying commenting lines

* Incorporating comments in scripts

* Modifying shell scripts for variable substitution

* Changing TLS version and cleanup of test scripts

* Fix shell script issue

* Modification of curl commands scripts

* Adding some changes in url of scripts

* Correcting indentation

* Adding missedout echo

+++ lte/backend/setup_env.sh
@@ -28,7 +28,7 @@ pkg_url=ftp://ftp.linux.ro/gentoo/distfiles/fcgi-2.4.1-SNAP-0910052249.tar.gz
 pkg_version=2.4.1-SNAP-0910052249
 pkg_name=fcgi-2.4.1-SNAP-0910052249
 log "Download $pkg_name"
-cd $setup_dir
+cd "$setup_dir"
 wget -c $pkg_url
 if [ $? -ne 0 ]; then
         log "$pkg_name package unavailable"
@@ -36,7 +36,7 @@ if [ $? -ne 0 ]; then
 fi
 log "Extract $pkg_name"
 tar -xvmf $pkg_name.tar.gz > /dev/null
-cd $setup_dir/$pkg_name
+cd "$setup_dir/$pkg_name"
 log "Build $pkg_name"
 
 sed -i 'N;24a#include <cstdio>' libfcgi/fcgio.cpp
@@ -47,7 +47,7 @@ if [ $? -ne 0 ]; then
         log "Compiled [ $pkg_name ] failed."
         exit 1
 fi
-rm -rf $setup_dir/$pkg_name*
+rm -rf "${setup_dir:?}/${pkg_name:?}"*
 
 
 # Install Jsoncpp
@@ -55,22 +55,22 @@ pkg_url=https://github.com/open-source-parsers/jsoncpp.git
 pkg_version=1.6.5
 pkg_name=jsoncpp
 log "Download $pkg_name"
-rm -rf $setup_dir/$pkg_name*
-cd $setup_dir
+rm -rf "${setup_dir:?}/${pkg_name:?}"*
+cd "$setup_dir"
 git clone -b $pkg_version $pkg_url
 if [ $? -ne 0 ]; then
         log "$pkg_name package unavailable"
         exit 1
 fi
-cd $setup_dir/$pkg_name
+cd "$setup_dir/$pkg_name"
 log "Build $pkg_name"
 cmake -H. -Bbuild && make -C build && make install -C build
 if [ $? -ne 0 ]; then
         log "Compiled [ $pkg_name ] failed."
         exit 1
 fi
-rm -rf $setup_dir/$pkg_name*
-cd $setup_dir
+rm -rf "${setup_dir:?}/${pkg_name:?}"*
+cd "$setup_dir"
 
 log "Done"
