+++ lte/http/install_nginx.sh
@@ -26,7 +26,7 @@ yum -y install pcre-devel.x86_64
 pkg_url=http://nginx.org/download/nginx-1.10.3.tar.gz
 pkg_name=nginx-1.10.3
 log "Download $pkg_name"
-cd $setup_dir
+cd "$setup_dir"
 wget -c $pkg_url
 if [ $? -ne 0 ]; then
         log "$pkg_name package unavailable"
@@ -34,7 +34,7 @@ if [ $? -ne 0 ]; then
 fi
 log "Extract $pkg_name"
 tar -xvmf $pkg_name.tar.gz > /dev/null
-cd $setup_dir/$pkg_name
+cd "$setup_dir"/"$pkg_name"
 log "Build $pkg_name"
 ./configure --with-http_ssl_module --with-http_v2_module && make install
 if [ $? -ne 0 ]; then
@@ -56,14 +56,13 @@ test -f /etc/nginx/nginx.conf && {
 }
 ln -s /usr/local/nginx/conf/nginx.conf /etc/nginx/
 
-rm -rf $setup_dir/$pkg_name*
+rm -rf "${setup_dir:?}/${pkg_name:?}"*
 
 #Install libfcgi
 pkg_url=ftp://ftp.linux.ro/gentoo/distfiles/fcgi-2.4.1-SNAP-0910052249.tar.gz
-pkg_version=2.4.1-SNAP-0910052249
 pkg_name=fcgi-2.4.1-SNAP-0910052249
 log "Download $pkg_name"
-cd $setup_dir
+cd "$setup_dir"
 wget -c $pkg_url
 if [ $? -ne 0 ]; then
         log "$pkg_name package unavailable"
@@ -71,7 +70,7 @@ if [ $? -ne 0 ]; then
 fi
 log "Extract $pkg_name"
 tar -xvmf $pkg_name.tar.gz > /dev/null
-cd $setup_dir/$pkg_name
+cd "$setup_dir"/"$pkg_name"
 log "Build $pkg_name"
 
 sed -i 'N;24a#include <cstdio>' libfcgi/fcgio.cpp
@@ -82,7 +81,7 @@ if [ $? -ne 0 ]; then
         log "Compiled [ $pkg_name ] failed."
         exit 1
 fi
-rm -rf $setup_dir/$pkg_name*
+rm -rf "${setup_dir:?}/${pkg_name:?}"*
 
 
 log "Done"
